Change asset serialization to text
This probably shouldn't be on this branch,
but we need to change asset serialization to
text so that things like the scene become mergable.

Otherwise, if two people work on the same file at
the same time, the second merge will nuke the first.

I've already had problems with this, so this should
make the process more pleasant.